% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subgroupsem_ready}
\alias{subgroupsem_ready}
\title{Checks whether python and module pysubgroup are installed.}
\usage{
subgroupsem_ready(ask = FALSE)
}
\arguments{
\item{ask}{Logical. Default is FALSE. Indicates whether user is asked whether
pysubgroup should be installed if it is not available.}
}
\description{
Checks whether python and module pysubgroup are installed.
}
