% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsem_userfun.R
\name{subsem}
\alias{subsem}
\title{User-level function with user-defined interestingness measure}
\usage{
subsem(model, data, qf = NULL, predictors = NULL, subsemOptions = NULL)
}
\arguments{
\item{model}{a lavaan model syntax (a character vector)}

\item{data}{a data frame}

\item{qf}{a lavaan syntax-based computation of the interestingness measure,
where the interestingness measure has to be named *subsem_qf*. Can also
be included directly in the model syntax (then, here NULL, as is default)}

\item{predictors}{a character vector of variable names, which are used as
covariates/predictors in the subgroup discovery (variables in data)}
}
\value{
List containing the time consumed and the groups.
}
\description{
todo
}
\examples{
# Define lavaan model
model <- "
eta1 =~ NA*x1 + c(la21,la22)*x2 + x3
eta2 =~ NA*x4 + c(la51,la52)*x5 + x6
eta3 =~ NA*x7 + c(la81,la82)*x8 + x9

eta1 ~~ 1*eta1
eta2 ~~ 1*eta2
eta3 ~~ 1*eta3

eta1 + eta2 + eta3 ~ 0*1

subsem_qf := abs(la21 - la22)
"

# Pass model, data and names of predictors to function
m1 <- subsem_wald(
  model = model,
  data = lavaan::HolzingerSwineford1939,
  qf = NULL,
  predictors = c("sex", "school", "grade")
)
summary(m1)
}
