% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsem_userfun.R
\name{subsem_lrt}
\alias{subsem_lrt}
\title{User-level function for LRT-based SubgroupSEM}
\usage{
subsem_lrt(model, data, predictors = NULL)
}
\arguments{
\item{model}{a lavaan model syntax (a character vector)}

\item{data}{a data frame}

\item{predictors}{a character vector of variable names, which are used as 
covariates/predictors in the subgroup discovery (variables in data)}
}
\description{
todo
}
