% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsem_userfun.R
\name{subsem_wald}
\alias{subsem_wald}
\title{User-level function for Wald-test based SubgroupSEM}
\usage{
subsem_wald(model, data, constraints, predictors = NULL)
}
\arguments{
\item{model}{a lavaan model syntax (a character vector)}

\item{data}{a data frame}

\item{constraints}{a lavaan syntax-based constraint of parameters for the 
Wald test. (a character vector)}

\item{predictors}{a character vector of variable names, which are used as 
covariates/predictors in the subgroup discovery (variables in data)}
}
\description{
todo
}
