% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsem_userfun.R
\name{subsem_wald}
\alias{subsem_wald}
\title{User-level function for Wald-test based SubgroupSEM}
\usage{
subsem_wald(model, data, constraints, predictors = NULL)
}
\arguments{
\item{model}{a lavaan model syntax (a character vector)}

\item{data}{a data frame}

\item{constraints}{a lavaan syntax-based constraint of parameters for the
Wald test. (a character vector)}

\item{predictors}{a character vector of variable names, which are used as
covariates/predictors in the subgroup discovery (variables in data)}
}
\value{
List containing the time consumed and the groups.
}
\description{
todo
}
\examples{
# Define lavaan model
model <- "
eta1 =~ NA*x1 + c(la21,la22)*x2 + x3
eta2 =~ NA*x4 + c(la51,la52)*x5 + x6
eta3 =~ NA*x7 + c(la81,la82)*x8 + x9

eta1 ~~ 1*eta1
eta2 ~~ 1*eta2
eta3 ~~ 1*eta3

eta1 + eta2 + eta3 ~ 0*1
"

con <- "
la21 == la22
la51 == la52
la81 == la82
"

# Pass model, data and names of predictors to function
m1 <- subsem_wald(
  model = model,
  data = lavaan::HolzingerSwineford1939,
  constraints = con,
  predictors = c("sex", "school", "grade")
)
summary(m1)
}
